requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-Login==0.6.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
packaging==24.2
Werkzeug==3.1.3
